migrate all functions to components
