+++ keccak256/src/permutation.rs
@@ -1,11 +1,7 @@
 #![allow(clippy::type_complexity)]
 #![allow(clippy::too_many_arguments)]
 pub mod circuit;
+pub(crate) mod components;
 pub(crate) mod generic;
-pub(crate) mod iota;
-pub(crate) mod pi;
-pub(crate) mod rho;
 pub(crate) mod rho_helpers;
 pub(crate) mod tables;
-pub(crate) mod theta;
-pub(crate) mod xi;
